osd: avoid breaking crimson builds
Fixes: https://tracker.ceph.com/issues/48182
